Moved classproperty from data.py to utils.py
